refactor: extract components and fix button flickering
