+++ src/renderer/components/AircraftSection/Enums.ts
@@ -1,0 +1,23 @@
+export enum InstallStatus {
+    UpToDate,
+    NeedsUpdate,
+    NotInstalled,
+    GitInstall,
+    TrackSwitch,
+    DownloadPrep,
+    Downloading,
+    Decompressing,
+    DownloadEnding,
+    DownloadDone,
+    DownloadRetry,
+    DownloadError,
+    DownloadCanceled,
+    Unknown,
+    Hidden,
+}
+
+export enum MsfsStatus {
+    Open,
+    Closed,
+    Checking,
+}
