+++ src/renderer/components/App/SideBar.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { Check, ChevronDown, Download, Refresh } from "tabler-icons-react";
 import { useSelector, } from "react-redux";
 import { InstallerStore } from "renderer/redux/store";
-import { InstallStatus } from "renderer/components/AircraftSection";
 import { Addon } from "renderer/utils/InstallerConfiguration";
+import { InstallStatus } from "renderer/components/AircraftSection/Enums";
 
 export type SidebarItemProps = { enabled?: boolean, selected: boolean, onClick: () => void, className?: string }
