+++ src/renderer/utils/AddonData.ts
@@ -3,13 +3,13 @@ import { GitVersions } from "@flybywiresim/api-client";
 import { Directories } from "./Directories";
 import fs from 'fs-extra';
 import { getCurrentInstall, needsUpdate } from "@flybywiresim/fragmenter";
-import { InstallStatus } from "renderer/components/AircraftSection";
 import settings from "common/settings";
 import { store, InstallerStore } from "renderer/redux/store";
 import { setInstalledTrack } from "renderer/redux/features/installedTrack";
 import { setSelectedTrack } from "renderer/redux/features/selectedTrack";
 import { setInstallStatus } from "renderer/redux/features/installStatus";
 import yaml from 'js-yaml';
+import { InstallStatus } from "renderer/components/AircraftSection/Enums";
 
 export type ReleaseInfo = {
     name: string,
